Move StaticsCache into a separate file.
