+++ app/lib/frontend/handlers.dart
@@ -10,7 +10,6 @@ import 'dart:math';
 
 import 'package:http_parser/http_parser.dart';
 import 'package:logging/logging.dart';
-import 'package:mime/mime.dart' as mime;
 import 'package:path/path.dart' as path;
 import 'package:shelf/shelf.dart' as shelf;
 
@@ -25,11 +24,15 @@ import 'backend.dart';
 import 'handlers_redirects.dart';
 import 'models.dart';
 import 'search_service.dart';
+import 'static_files.dart';
 import 'templates.dart';
 
 final _pubHeaderLogger = new Logger('pub.header_logger');
 
 final String _staticPath = Platform.script.resolve('../../static').toFilePath();
+
+/// Stores binary data of /static
+final StaticsCache staticsCache = new StaticsCache(_staticPath);
 
 // Non-revealing metrics to monitor the search service behavior from outside.
 final _packageAnalysisLatencyTracker = new LastNTracker<Duration>();
@@ -220,9 +223,6 @@ Future<shelf.Response> packagesHandler(shelf.Request request) async {
     return packagesHandlerHtml(request);
   }
 }
-
-/// Handles requests for /static
-final StaticsCache staticsCache = new StaticsCache();
 
 /// The max age a browser would take hold of the static files before checking
 /// with the server for a newer version.
@@ -629,33 +629,6 @@ int _pageFromUrl(Uri url) {
     } catch (_, __) {}
   }
   return pageAsInt;
-}
-
-class StaticsCache {
-  final Map<String, StaticFile> staticFiles = <String, StaticFile>{};
-
-  StaticsCache() {
-    final Directory staticsDirectory = new Directory(_staticPath);
-    final files = staticsDirectory
-        .listSync(recursive: true)
-        .where((fse) => fse is File)
-        .map((file) => file.absolute);
-
-    for (final File file in files) {
-      final contentType = mime.lookupMimeType(file.path) ?? 'octet/binary';
-      final bytes = file.readAsBytesSync();
-      final lastModified = file.lastModifiedSync();
-      staticFiles[file.path] = new StaticFile(contentType, bytes, lastModified);
-    }
-  }
-}
-
-class StaticFile {
-  final String contentType;
-  final List<int> bytes;
-  final DateTime lastModified;
-
-  StaticFile(this.contentType, this.bytes, this.lastModified);
 }
 
 bool _isProd(shelf.Request request) {
